backend/requirements.txt
fastapi==0.104.0
requests==2.31.0
python-dotenv==1.0.0
uvicorn==0.23.2
pandas
datasets==2.14.4
sentence-transformers==2.2.2
numpy==1.24.3
pydantic==2.4.2